fix output format (from this [area, area] to this [[area], [area]]) and refactoring
