+++ nodes/analyzer/volume.py
@@ -17,20 +17,18 @@
 # ##### END GPL LICENSE BLOCK #####
 
 import bpy
-import bmesh
 from sverchok.node_tree import SverchCustomTreeNode
 from sverchok.utils.sv_bmesh_utils import bmesh_from_pydata
-from sverchok.data_structure import dataCorrect, Vector_generate
 
 
-class SvVolumeNode(bpy.types.Node, SverchCustomTreeNode):
-    ''' Volume '''
-    bl_idname = 'SvVolumeNode'
+class SvVolumeNodeMK2(SverchCustomTreeNode, bpy.types.Node):
+    """
+    Triggers: volume
+    Tooltip: Calculate volume of an object
+    """
+    bl_idname = 'SvVolumeNodeMK2'
     bl_label = 'Volume'
     bl_icon = 'SNAP_VOLUME'
-
-    def draw_buttons(self, context, layout):
-        pass
 
     def sv_init(self, context):
         self.inputs.new('SvVerticesSocket', 'Vers')
@@ -38,49 +36,22 @@ class SvVolumeNode(bpy.types.Node, SverchCustomTreeNode):
         self.outputs.new('SvStringsSocket', "Volume")
 
     def process(self):
-        verts_socket, polys_socket = self.inputs
-        vol_socket = self.outputs[0]
+        vertices = self.inputs['Vers'].sv_get(deepcopy=False, default=[])
+        faces = self.inputs['Pols'].sv_get(deepcopy=False, default=[])
 
-        if vol_socket.is_linked and verts_socket.is_linked:  # and polys_socket.is_linked ?
+        out = []
+        for verts_obj, faces_obj in zip(vertices, faces):
+            bm = bmesh_from_pydata(verts_obj, [], faces_obj, normal_update=True)
+            out.append([bm.calc_volume()])
+            bm.free()
 
-            vertices = Vector_generate(dataCorrect(verts_socket.sv_get()))
-            faces = dataCorrect(polys_socket.sv_get())
+        self.outputs[0].sv_set(out)
 
-            out = []
-            for verts_obj, faces_obj in zip(vertices, faces):
-                bm = bmesh_from_pydata(verts_obj, [], faces_obj, normal_update=True)
-                out.append(bm.calc_volume())
-                bm.free()
- 
-            vol_socket.sv_set(out)
-
-    '''
-    solution, that blow my mind, not delete.
-    i have to investigate it here
-    
-    def Volume(self, bme):
-        verts = obj_data.vertices     # array of vertices
-        obj_data.calc_tessface()
-        faces = obj_data.tessfaces        # array of faces
-        VOLUME = 0;     # VOLUME OF THE OBJECT
-        
-        for f in faces:
-            fverts = f.vertices      # getting face's vertices
-            ab = verts[fverts[0]].co 
-            ac = verts[fverts[1]].co
-            ad = verts[fverts[2]].co
-            
-            # calculating determinator
-            det = (ab[0]*ac[1]*ad[2]) - (ab[0]*ac[2]*ad[1]) - \
-                (ab[1]*ac[0]*ad[2]) + (ab[1]*ac[2]*ad[0]) + \
-                (ab[2]*ac[0]*ad[1]) - (ab[2]*ac[1]*ad[0])
-            
-            VOLUME += det/6
-    '''
 
 def register():
-    bpy.utils.register_class(SvVolumeNode)
+    bpy.utils.register_class(SvVolumeNodeMK2)
 
 
 def unregister():
-    bpy.utils.unregister_class(SvVolumeNode)
+    bpy.utils.unregister_class(SvVolumeNodeMK2)
+
